requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-extensions==3.2.3
django-oauth-toolkit==2.1.0
django-phonenumber-field==7.1.0
django-sslserver==0.22
oauthlib==3.2.0
phonenumbers==8.13.14
drf_yasg==1.21.6
pydub==0.25.1
python-dotenv==1.0.0
pyVoIP==1.6.4
pyVoIP==1.6.4
Requests==2.31.0
